[Refactor] Add `PerformanceCalculator` class and move `calculatePerformanceAmount` into it.
My idea with this is that I want to separate the code responsible for doing calculations and the code responsible for formatting and presenting the statement.

I'm also thinking that I should get rid of the switch statement and use Polymorphism to do it but I haven't figured out how just yet.