[Refactor] Extract statement text production from `print` function into it's own `produceStatementText` function
Finally gotten the `print` function to a level of abstraction close to what I was looking for. Any concepts of calculation are now separated from the string building.

It's not quite right yet though. I don't like that I'm passing in `invoice` along with `totalAmount` and `volumeCredits` as these parameters are not at the same level of abstraction.

Stay tuned to see how I fix this in the next commits.