[Refactor] Include `customer` and `performances` in `Statement` interface
This removes the need to pass `invoice` to `produceStatementText`. The only thing that remains out of place is the fact that we are still using the `performanceCalculator` in the `produceStatementText` function.

It feels like this calculation should be happening in `generateStatement` in the data passed along in the `Statement` interface. That's the next step, I guess.